+++ pephubclient/exceptions.py
@@ -31,3 +31,11 @@ class PEPExistsError(BasePephubclientException):
     def __init__(self, message: Optional[str] = None):
         self.message = message
         super().__init__(self.message or self.default_message)
+
+
+class FileDoesNotExistError(BasePephubclientException):
+    default_message = "File does not exist."
+
+    def __init__(self, message: Optional[str] = None):
+        self.message = message
+        super().__init__(self.message or self.default_message)